finsec-api/finsec_api/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
PyJWT==2.1.0
pymysql==1.0.2
cryptography==36.0.0
python-dotenv==0.19.1
Werkzeug==2.0.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
mysqlclient==2.1.0
pyotp==2.6.0
flask-swagger-ui==4.11.1
flasgger==0.9.7.1
unidecode==1.3.6 